refactor: 회원가입 로직 리팩터링 및 응답 구조 개선
- 컨트롤러의 이메일 중복 검사 로직을 서비스 계층으로 이동
- 회원가입 응답 구조 변경
- 회원가입 메서드면 signupUser -> addUser로 변경하여 컨벤션 통일

S12P21A304-95

+++ backend/src/main/java/newstock/domain/user/repository/UserCustomRepositoryImpl.java
@@ -19,7 +19,7 @@ public class UserCustomRepositoryImpl implements UserCustomRepository {
     }
 
     @Override
-    public Optional<User> findById(Long userId) {
+    public Optional<User> findById(Integer userId) {
         return Optional.ofNullable(jpaQueryFactory
                 .selectFrom(user)
                 .where(user.userId.eq(userId))
